requirements.txt
requests==2.31.0
allure-pytest==2.13.2
pytest==7.4.0
selenium==4.11.2
webdriver-manager==4.0.0
pysftp==0.2.9